server/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
flask-socketio
python-dotenv
Pillow
werkzeug
openai
socketio
gunicorn 
eventlet
google-cloud-speech
pyaudio
webrtcvad
soundfile
numpy
pydub
ffmpeg-python
elevenlabs
oauthlib
requests
requests_oauthlib==1.3.1
google-auth
aiohttp
tiktoken
stripe>=11.0.0
cryptography
aiohttp 
ratelimit
psycopg2-binary
PyMySQL
boto3 
asyncio
Flask[async]
apscheduler
httpagentparser